finding start and end using warp communication

+++ device/cuda/src/clusterization/clusterization_algorithm.cu
@@ -490,7 +490,7 @@ __global__ void ccl_kernel2(
     */
    __shared__ short flag[2];  
    unsigned int short cell = 999; 
-   __shared__ int short minWho[4];
+   __shared__ unsigned int  minWho[4];
 
 
     #pragma unroll
